allowed users to donate floating decimal not just whole numbers

+++ client/src/utils/queries.js
@@ -113,7 +113,7 @@ query CountByCategory {
 `;
 
 export const QUERY_DONATE = gql`
-  query donate($donationAmount: Int!) {
+  query donate($donationAmount: Float!) {
     donate(donationAmount: $donationAmount) {
       session
     }
